+++ src/riemopt.py
@@ -4,21 +4,12 @@ import jax
 import jax.config
 jax.config.update("jax_enable_x64", True)
 
-from matrix import TuckerMatrix
-from tucker import Tucker
+from src.matrix import TuckerMatrix
+from src.tucker import Tucker
 import numpy as np
 import jax.numpy as jnp
 
-A = np.diag([1, 2, 3, 4])
-Q, _ = np.linalg.qr(np.random.random((4, 4)))
-A = Q @ A @ Q.T
-A = TuckerMatrix.full2tuck(A.reshape([2] * 4), [2] * 2, [2] * 2)
 
-@jax.jit
-def f(T):
-    return (T.flat_inner(A @ T)) / T.flat_inner(T)
-
-@jax.jit
 def group_cores(core1, core2):
     d = len(core1.shape)
     r = core1.shape
@@ -32,20 +23,8 @@ def group_cores(core1, core2):
 
     return new_core
 
-@jax.jit
-def g(T1, core, factors):
-    d = T1.ndim
-    r = T1.rank
 
-    new_factors = [jnp.concatenate([T1.factors[i], factors[i]], axis=1) for i in range(T1.ndim)]
-    new_core = group_cores(core, T1.core)
-
-    T = Tucker(new_core, new_factors)
-    return f(T)
-
-
-@jax.jit
-def compute_gradient_projection(T):
+def compute_gradient_projection(T, g):
     """
     Input
         X: tensor from manifold
@@ -62,7 +41,7 @@ def compute_gradient_projection(T):
     return Tucker(group_cores(dS, T.core), [jnp.concatenate([T.factors[i], dU[i]], axis=1) for i in range(T.ndim)])
 
 
-def optimize(f, X0, maxiter=10):
+def optimize(f, g, X0, maxiter=10):
     """
     Input
         f: function to maximize
@@ -81,7 +60,7 @@ def optimize(f, X0, maxiter=10):
     
     for i in range(maxiter):
         print(f'Doing iteration {i+1}/{maxiter}\t Calculating gradient...\t', end='\r')
-        G = compute_gradient_projection(X)
+        G = compute_gradient_projection(X, g)
         print(f'Doing itaration {i+1}/{maxiter}\t Calculating tau...\t\t', end='\r')
         tau = 1
         print(f'Doing iteration {i+1}/{maxiter}\t Calculating retraction...\t', end='\r')
@@ -92,17 +71,3 @@ def optimize(f, X0, maxiter=10):
         print(f'Done iteration {i+1}/{maxiter}!\t Error: {errs[-1]}' + ' ' * 50, end='\r')
         
     return X, errs
-
-np.random.seed(229)
-
-x = np.random.random(4)
-X = Tucker.full2tuck(x.reshape([2] * 2))
-print(f(X))
-
-X, _ = optimize(f, X, maxiter=10)
-x = X.full().reshape(4)
-A = A.full().reshape((4, 4))
-
-print(x)
-print(A @ x)
-print((x.T @ A @ x) / (x.T @ x))
